+++ src/document/DjVuModel.h
@@ -45,7 +45,7 @@ public:
 
     QSizeF sizeF() const;
 
-    QImage render(int width, int height, Qt::AspectRatioMode mode)const;     //按大小获取
+    QImage render(qreal width, qreal height, Qt::AspectRatioMode mode)const;     //按大小获取
 
     QImage render(Dr::Rotation rotation, const double scaleFactor, const QRectF &boundingRect = QRectF()) const;//按缩放比例获取; boundingRect:取其中某一区域图片
